% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/many_sample_means.R
\name{many_sample_means}
\alias{many_sample_means}
\title{Generating the means of a sub-sample from a vector a specified amount of times
and then outputting it into a vector}
\usage{
many_sample_means(vec, n, reps)
}
\arguments{
\item{vec}{a numeric vector}

\item{n}{an integer that indicates the random sample size}

\item{reps}{an integer that denotes the number of repetitions to be conducted}
}
\value{
a vector of the sample means
}
\description{
Generating the means of a sub-sample from a vector a specified amount of times
and then outputting it into a vector
}
