% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_mean.R
\name{sample_mean}
\alias{sample_mean}
\title{Generating the means of a sub-sample from a vector and then outputting it into a vector}
\usage{
sample_mean(vec, n)
}
\arguments{
\item{vec}{a numeric vector}

\item{n}{an integer that indicates the random sample size}
}
\value{
a vector of sample means
}
\description{
Generating the means of a sub-sample from a vector and then outputting it into a vector
}
