% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_means_ns.R
\name{sample_means_ns}
\alias{sample_means_ns}
\title{Generating the means of a number of sub-samples from a vector and then outputting it into a data frame}
\usage{
sample_means_ns(vec, reps, ns)
}
\arguments{
\item{vec}{a numeric vector}

\item{reps}{an integer that denotes the number of repetitions to be conducted}

\item{ns}{an vector that indicates the desired random sample sizes}
}
\value{
a data frame of the sample means
}
\description{
Generating the means of a number of sub-samples from a vector and then outputting it into a data frame
}
