% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/many_sample_means.R
\name{many_sample_means}
\alias{many_sample_means}
\title{Takes a vector and from it takes a random sample of size n then calculates sample
mean and repeats it reps times}
\usage{
many_sample_means(vec, reps, n)
}
\arguments{
\item{vec}{A vector}

\item{reps}{A number indicating the number of times to repeat sample mean process}

\item{n}{A number to represent the size of the sample}
}
\value{
A vector containing the sample means.
}
\description{
Takes a vector and from it takes a random sample of size n then calculates sample
mean and repeats it reps times
}
