% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_mean.R
\name{sample_mean}
\alias{sample_mean}
\title{Takes a vector and from it takes a random sample of size n, then calculates sample mean}
\usage{
sample_mean(vec, n)
}
\arguments{
\item{vec}{A vector}

\item{n}{A number to represent the size of the sample}
}
\value{
A vector containing the sample mean.
}
\description{
Takes a vector and from it takes a random sample of size n, then calculates sample mean
}
