% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_means_ns.R
\name{sample_means_ns}
\alias{sample_means_ns}
\title{Takes a vector and from it takes a random sample of size n (each value in ns)
then calculates sample mean and repeats it reps times}
\usage{
sample_means_ns(vec, reps, ns)
}
\arguments{
\item{vec}{A vector}

\item{reps}{A number indicating the number of times to repeat sample mean process}

\item{ns}{A vector of numbers to represent the size of the sample}
}
\value{
A data frame with two columns, one for the sample mean and one for the associated sample size.
}
\description{
Takes a vector and from it takes a random sample of size n (each value in ns)
then calculates sample mean and repeats it reps times
}
